add client

+++ server/routes/lists.js
@@ -1,11 +1,11 @@
 const router = require("express").Router();
 const List = require("../models/List");
-// const verify = require("../verifyToken");
+const verify = require("../verifyToken");
 
 //CREATE
 
 router.post("/", async (req, res) => {
-  // if (req.user.isAdmin) {
+  if (req.user.isAdmin) {
     const newList = new List(req.body);
     try {
       const savedList = await newList.save();
@@ -13,24 +13,24 @@ router.post("/", async (req, res) => {
     } catch (err) {
       res.status(500).json(err);
     }
-  // } else {
-  //   res.status(403).json("You are not allowed!");
-  // }
+  } else {
+    res.status(403).json("You are not allowed!");
+  }
 });
 
 //DELETE
 
 router.delete("/:id", async (req, res) => {
-  // if (req.user.isAdmin) {
+  if (req.user.isAdmin) {
     try {
       await List.findByIdAndDelete(req.params.id);
       res.status(201).json("The list has been delete...");
     } catch (err) {
       res.status(500).json(err);
     }
-  // } else {
-  //   res.status(403).json("You are not allowed!");
-  // }
+  } else {
+    res.status(403).json("You are not allowed!");
+  }
 });
 
 //GET
